+++ libs/model/src/services/commonProtocol/contractHelpers.ts
@@ -6,7 +6,7 @@ import {
 } from '@hicommonwealth/shared';
 import Web3 from 'web3';
 import { AbiItem } from 'web3-utils';
-import { config } from '../../config';
+
 import { Balances, TokenAttributes, getBalances } from '../tokenBalanceCache';
 import { contestABI } from './abi/contestAbi';
 
@@ -140,20 +140,4 @@ export const getTokenAttributes = async (
     ticker: String(symbol),
     decimals: parseInt(String(decimals)),
   };
-};
-
-/**
- * A helper for creating the web3 provider via an RPC, including private key import
- * @param rpc the rpc of the network to use helper with
- * @returns
- */
-export const createWeb3Provider = async (rpc: string): Promise<Web3> => {
-  if (!config.WEB3.PRIVATE_KEY) throw new AppError('WEB3 private key not set!');
-  const web3 = new Web3(rpc);
-  const account = web3.eth.accounts.privateKeyToAccount(
-    config.WEB3.PRIVATE_KEY,
-  );
-  web3.eth.accounts.wallet.add(account);
-  web3.eth.defaultAccount = account.address;
-  return web3;
 };
