+++ libs/model/src/services/commonProtocol/index.ts
@@ -6,9 +6,8 @@ import * as contest from './contestHelper';
 import * as contract from './contractHelpers';
 import * as launchpad from './launchpadHelpers';
 
-// export modules as objects so they can be stubbed in tests
-export const communityStakeConfigValidator = stake;
-export const contractHelpers = contract;
-export const contestHelper = contest;
-export const aaWallet = wallet;
-export const launchpadHelpers = launchpad;
+export const communityStakeConfigValidator = { ...stake };
+export const contractHelpers = { ...contract };
+export const contestHelper = { ...contest };
+export const aaWallet = { ...wallet };
+export const launchpadHelpers = { ...launchpad };
